Clean up Package and Mapping entities
