+++ src/presentation/webapp/components/mapping-dialog/MappingDialog.tsx
@@ -6,8 +6,9 @@ import { ConfirmationDialog, OrgUnitsSelector } from "d2-ui-components";
 import _ from "lodash";
 import React, { useEffect, useState } from "react";
 import { Instance as InstanceEntity } from "../../../../domain/instance/entities/Instance";
+import { MetadataMappingDictionary } from "../../../../domain/instance/entities/MetadataMapping";
 import { d2ModelFactory } from "../../../../models/dhis/factory";
-import Instance, { MetadataMappingDictionary } from "../../../../models/instance";
+import Instance from "../../../../models/instance";
 import { D2 } from "../../../../types/d2";
 import { MetadataType } from "../../../../utils/d2";
 import { useAppContext } from "../../../common/contexts/AppContext";
